+++ src/main/kotlin/ghidra/app/util/bin/format/elf/PspElfConstants.kt
@@ -15,4 +15,5 @@ object PspElfConstants {
   val ET_PSP_PRX = 0xFFA0.toShort()
 
   val SHT_PSP_REL = 0x700000A0
+  val SHT_PSP_REL_TYPE_B = 0x700000A1
 }