+++ src/main/kotlin/ghidra/app/util/opinion/PspElfLoader.kt
@@ -1,19 +1,39 @@
 package ghidra.app.util.opinion
 
 import ghidra.app.util.Option
+import ghidra.app.util.OptionUtils
 import ghidra.app.util.bin.ByteProvider
 import ghidra.app.util.bin.format.elf.ElfException
 import ghidra.app.util.bin.format.elf.PspElfHeader
 import ghidra.app.util.importer.MessageLog
 import ghidra.app.util.importer.MessageLogContinuesFactory
+import ghidra.framework.model.DomainObject
 import ghidra.program.model.listing.Program
-import ghidra.util.exception.CancelledException
 import ghidra.util.task.TaskMonitor
 import java.io.IOException
 
 class PspElfLoader : ElfLoader() {
   companion object {
     const val PSP_ELF_NAME = "PSP Executable (ELF)"
+
+    object Options {
+      private const val COMMAND_LINE_ARG_PREFIX = "-psp"
+
+      object UseRebootBinTypeBMapping {
+        private const val NAME = "Use reboot.bin Type B Relocation Mapping"
+        private const val DEFAULT = false
+        private val TYPE = java.lang.Boolean::class.java // must be Java Boolean
+        private const val COMMAND_LINE = "${COMMAND_LINE_ARG_PREFIX}-useRebootBinTypeBRelocationMapping"
+
+        fun toOption(): Option {
+          return Option(NAME, DEFAULT, TYPE, COMMAND_LINE)
+        }
+
+        fun getValue(options: List<Option>): Boolean {
+          return OptionUtils.getOption(NAME, options, DEFAULT)
+        }
+      }
+    }
   }
 
   override fun getName(): String {
@@ -28,18 +48,27 @@ class PspElfLoader : ElfLoader() {
     return 0
   }
 
+  override fun getDefaultOptions(
+    provider: ByteProvider?,
+    loadSpec: LoadSpec?,
+    domainObject: DomainObject?,
+    loadIntoProgram: Boolean
+  ): MutableList<Option> {
+    val options = super.getDefaultOptions(provider, loadSpec, domainObject, loadIntoProgram)
+    options.add(Options.UseRebootBinTypeBMapping.toOption())
+    return options
+  }
+
   override fun load(
     provider: ByteProvider, loadSpec: LoadSpec?, options: List<Option>,
     program: Program, monitor: TaskMonitor, log: MessageLog
   ) {
     try {
       val factory = MessageLogContinuesFactory.create(log)
-      val elf = PspElfHeader.createElfHeader(factory, provider)
+      val elf = PspElfHeader.createElfHeader(factory, provider, Options.UseRebootBinTypeBMapping.getValue(options))
       ElfProgramBuilder.loadElf(elf, program, options, log, monitor)
       program.executableFormat = PSP_ELF_NAME
     } catch (e: ElfException) {
-      throw IOException(e.message)
-    } catch (e: CancelledException) { // TODO: Caller should properly handle CancelledException instead
       throw IOException(e.message)
     }
   }